A series of incorrect tests for the semantics.
These won't generate any C code, instead the generation of C code will
lead to an error because there was an error in the execution.

The aim with these tests is to verify that the border cases like
overflow or semantic errors that the syntax allows are in fact an error
in our semantics.

If one of these tests starts generating code then most likely there's
something wrong in the latest change to the semantics since none of these
test cases should generate any code and should reach an execution error.